wo#6269 . split out reading of /proc/net/pf_key_supported
