+++ lib/libalgoparse/Makefile
@@ -32,7 +32,7 @@ MANDIR=$(MANTREE)/man3
 ONEFILE=alg_info.c
 SRCS=alg_info.c
 SRCS+=plutoalg.c ipsecalg.c kernel_alg.c
-SRCS+=
+SRCS+=klips_interface.c
 
 #
 CFLAGS+=${IKE_CRYPTO_DEFINES}
